requirements.txt
flask
faiss-cpu
sentence-transformers
pandas
matplotlib
seaborn
requests
